tests/test_PyFileExecutor.py
"""
### Explanation:
1. **Test Setup and Teardown**:
   - `setUp` method initializes a temporary directory and constructs a file path for the script.
   - `tearDown` method ensures the temporary directory is cleaned up after tests are run.

2. **Mocks**:
   - `subprocess.run` is mocked to simulate the execution of the Python script without actually running it.
   - `os.makedirs` is mocked to test directory creation logic without affecting the real filesystem.

3. **Test Cases**:
   - **`test_execute_without_error`**: Tests the execution of the script when there is no error (`returncode = 0`).
   - **`test_execute_with_error`**: Tests the execution of the script when there is an error (`returncode = 1`).
   - **`test_execute_with_directory`**: Tests the script execution and directory creation logic.

The tests verify:
- The `subprocess.run` method is called with the correct command.
- The directory creation logic is correctly called.
- The error output is correctly captured if any.
- The executed code matches the input code.
"""

import unittest
import os
import tempfile
from unittest.mock import patch, Mock
from AutoChatBot.PyFileExecutor import PyFileExecutor

class TestPyFileExecutor(unittest.TestCase):
    
    def setUp(self):
        self.temp_dir = tempfile.TemporaryDirectory()
        self.file_path = os.path.join(self.temp_dir.name, 'script.py')

    def tearDown(self):
        self.temp_dir.cleanup()

    @patch('subprocess.run')
    def test_execute_without_error(self, mock_subprocess_run):
        self.code = "print('Hello, World!')"
        mock_subprocess_run.return_value = Mock(returncode=0, stderr=b'')
        
        executor = PyFileExecutor(self.file_path, self.code)
        executed_code, error_output = executor.execute()

        self.assertEqual(mock_subprocess_run.call_count, 1)
        expected_command = ['python', self.file_path]
        mock_subprocess_run.assert_called_with(expected_command, capture_output=True)

        self.assertIsNone(error_output)
        self.assertEqual(executed_code, self.code)

    @patch('subprocess.run')
    def test_execute_with_error(self, mock_subprocess_run):
        self.code = "print('Hello, World!')"
        mock_subprocess_run.return_value = Mock(returncode=1, stderr=b'Error: Division by zero')
        
        executor = PyFileExecutor(self.file_path, self.code)
        executed_code, error_output = executor.execute()

        self.assertEqual(mock_subprocess_run.call_count, 1)
        expected_command = ['python', self.file_path]
        mock_subprocess_run.assert_called_with(expected_command, capture_output=True)

        self.assertEqual(error_output, 'Error: Division by zero')
        self.assertEqual(executed_code, self.code)

    @patch('os.makedirs')
    @patch('subprocess.run')
    def test_execute_with_directory(self, mock_subprocess_run, mock_makedirs):
        self.code = "print('Hello, World!')"
        mock_subprocess_run.return_value = Mock(returncode=0, stderr=b'')

        directory = os.path.dirname(self.file_path)
        executor = PyFileExecutor(self.file_path, self.code)
        executed_code, error_output = executor.execute()
        
        self.assertEqual(mock_makedirs.call_count, 1)
        mock_makedirs.assert_called_with(directory, exist_ok=True)

        self.assertEqual(mock_subprocess_run.call_count, 1)
        expected_command = ['python', self.file_path]
        mock_subprocess_run.assert_called_with(expected_command, capture_output=True)

        self.assertIsNone(error_output)
        self.assertEqual(executed_code, self.code)

if __name__ == '__main__':
    unittest.main()
